refactor(utils): split type utils into modules
