+++ packages/lumx-react/src/utils/type/Callback.ts
@@ -1,0 +1,4 @@
+/**
+ * Callback function type alias (use for readability)
+ */
+export type Callback = () => void;
