+++ packages/lumx-react/src/utils/type/Falsy.ts
@@ -1,0 +1,5 @@
+/**
+ * JS falsy values.
+ * (excluding `NaN` as it can't be distinguished from `number`)
+ */
+export type Falsy = false | undefined | null | 0 | '';
