+++ packages/lumx-react/src/utils/type/HasClassName.ts
@@ -1,0 +1,6 @@
+export interface HasClassName {
+    /**
+     * Class name forwarded to the root element of the component.
+     */
+    className?: string;
+}
