+++ packages/lumx-react/src/utils/type/HasPolymorphicAs.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export type HasPolymorphicAs<E extends React.ElementType> = React.ComponentPropsWithoutRef<E> & {
+    /**
+     * Customize the rendered component.
+     */
+    as?: E;
+};
