+++ packages/lumx-react/src/utils/type/HasTheme.ts
@@ -1,0 +1,8 @@
+import type { Theme } from '@lumx/react';
+
+export interface HasTheme {
+    /**
+     * Theme adapting the component to light or dark background.
+     */
+    theme?: Theme;
+}
