+++ packages/lumx-react/src/utils/type/HeadingElement.ts
@@ -1,0 +1,2 @@
+/** Union type of all heading elements */
+export type HeadingElement = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
