+++ packages/lumx-react/src/utils/type/Point.ts
@@ -1,0 +1,4 @@
+/**
+ * A point coordinate in 2D space
+ */
+export type Point = { x: number; y: number };
