+++ packages/lumx-react/src/utils/type/Predicate.ts
@@ -1,0 +1,2 @@
+/** Predicate function type */
+export type Predicate<T> = (t: T) => boolean;
