+++ packages/lumx-react/src/utils/type/RectSize.ts
@@ -1,0 +1,4 @@
+/**
+ * Rectangle size
+ */
+export type RectSize = { width: number; height: number };
