+++ packages/lumx-react/src/utils/type/TextElement.ts
@@ -1,0 +1,4 @@
+import type { HeadingElement } from './HeadingElement';
+
+/** Union type of all text elements */
+export type TextElement = 'span' | 'p' | HeadingElement;
