+++ packages/lumx-react/src/utils/type/ValueOf.ts
@@ -1,0 +1,2 @@
+/** Get types of the values of a record. */
+export type ValueOf<T extends Record<any, any>> = T[keyof T];
